+++ SerenityLib/examples/solarSystem/game/gui/Widget.h
@@ -8,31 +8,31 @@
 class Widget : public IWidget {
     protected:
         void*                             m_UserPointer;
-        float                             m_Width;
-        float                             m_Height;
-        glm::vec2                         m_Position;
+        unsigned int                      m_Width;
+        unsigned int                      m_Height;
+        glm::uvec2                        m_Position;
         glm::vec4                         m_Color;
         bool                              m_Hidden;
         bool                              m_MouseIsOver;
         Alignment::Type                   m_Alignment;
 
-        const glm::vec2 positionFromAlignment();
-        const glm::vec2 positionFromAlignment(const float& width, const float& height, const Alignment::Type& alignment);
+        const glm::uvec2 positionFromAlignment();
+        const glm::uvec2 positionFromAlignment(const unsigned int& width, const unsigned int& height, const Alignment::Type& alignment);
     public:
-        Widget(const glm::vec2& position, const float& width, const float& height);
-        Widget(const float& x, const float& y, const float& width, const float& height);
+        Widget(const glm::vec2& position, const unsigned int& width, const unsigned int& height);
+        Widget(const unsigned int& x, const unsigned int& y, const unsigned int& width, const unsigned int& height);
         virtual ~Widget();
 
-        virtual void setWidth(const float&);
-        virtual void setHeight(const float&);
-        virtual void setSize(const float& width, const float& height);
+        virtual void setWidth(const unsigned int&);
+        virtual void setHeight(const unsigned int&);
+        virtual void setSize(const unsigned int& width, const unsigned int& height);
 
-        virtual const float width() const;
-        virtual const float height() const;
+        virtual const unsigned int width() const;
+        virtual const unsigned int height() const;
 
         virtual const bool isMouseOver() const;
 
-        const glm::vec2& position() const;
+        const glm::uvec2& position() const;
         const glm::vec4& color() const;
 
         void setAlignment(const Alignment::Type& alignment);
@@ -40,8 +40,8 @@ class Widget : public IWidget {
         void* getUserPointer() const;
         void setUserPointer(void*);
 
-        virtual void setPosition(const float& x, const float& y);
-        virtual void setPosition(const glm::vec2& position);
+        virtual void setPosition(const unsigned int& x, const unsigned int& y);
+        virtual void setPosition(const glm::uvec2& position);
 
         virtual void setColor(const float& r, const float& g, const float& b, const float& a);
         virtual void setColor(const glm::vec4& color);
